+++ DawnLib/src/API/Moons/DawnMoonInfo.cs
@@ -10,9 +10,11 @@ namespace Dawn;
 
 public sealed class DawnMoonInfo : DawnBaseInfo<DawnMoonInfo>
 {
-    internal DawnMoonInfo(NamespacedKey<DawnMoonInfo> key, HashSet<NamespacedKey> tags, SelectableLevel level, float outsideEnemiesProbabilityRange, int maxDaytimeDiversityPowerCount, int maxWeedEnemyPowerCount, int maxWeedDiversityPowerCount, List<SpawnableEnemyWithRarity> weedEnemies, AnimationCurve weedEnemySpawnChanceThroughDay, float weedEnemiesProbabilityRange, List<IMoonSceneInfo> scenes, TerminalNode? infoNode, TerminalNode? routeNode, TerminalNode? receiptNode, TerminalKeyword? nameKeyword, DawnPurchaseInfo dawnPurchaseInfo, IDataContainer? customData) : base(key, tags, customData)
+    internal DawnMoonInfo(NamespacedKey<DawnMoonInfo> key, HashSet<NamespacedKey> tags, SelectableLevel level, DawnTimeOfDayCues timeOfDayCues, float outsideEnemiesProbabilityRange, int maxDaytimeDiversityPowerCount, int maxWeedEnemyPowerCount, int maxWeedDiversityPowerCount, List<SpawnableEnemyWithRarity> weedEnemies, AnimationCurve weedEnemySpawnChanceThroughDay, float weedEnemiesProbabilityRange, List<IMoonSceneInfo> scenes, TerminalNode? infoNode, TerminalNode? routeNode, TerminalNode? receiptNode, TerminalKeyword? nameKeyword, DawnPurchaseInfo dawnPurchaseInfo, IDataContainer? customData) : base(key, tags, customData)
     {
         Level = level;
+
+        TimeOfDayCues = timeOfDayCues;
 
         OutsideEnemiesProbabilityRange = outsideEnemiesProbabilityRange;
 
@@ -35,6 +37,9 @@ public sealed class DawnMoonInfo : DawnBaseInfo<DawnMoonInfo>
     }
 
     public SelectableLevel Level { get; }
+
+    public DawnTimeOfDayCues TimeOfDayCues { get; private set; }
+
     public TerminalNode? InfoNode { get; }
     public TerminalNode? RouteNode { get; }
     public TerminalNode? ReceiptNode { get; }